requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.46
beautifulsoup4==4.3.2
characteristic==0.1.0
chardet==2.3.0
colorama==0.3.2
command-not-found==0.3
debtagshw==0.1
decorator==4.0.9
defer==1.0.6
dirspec==13.10
Django==1.8.6
django-crispy-forms==1.6.0
django-oauth2-provider==0.2.6.1
django-shell-plus==1.1.5
django-suit==0.2.18
django-tastypie==0.12.2
duplicity==0.7.1
gyp==0.1
html5lib==0.999
httplib2==0.9
idna==0.9
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
lockfile==0.8
lxml==3.4.2
ndg-httpsclient==0.3.2
oauthlib==0.6.1
oneconf==0.3.7
PAM==0.4.2
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==2.7.0
piston-mini-client==0.7.5
ptyprocess==0.5.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycrypto==2.6.1
pycups==1.9.72
pygobject==3.14.0
pyOpenSSL==0.13.1
pyserial==2.6
python-apt===0.9.3.11build1
python-dateutil==2.4.2
python-debian==0.1.22
python-mimeparse==0.1.4
pytz==2014.10
pyxdg==0.25
requests==2.4.3
service-identity==1.0.0
sessioninstaller==0.0.0
shortuuid==0.4.2
simplegeneric==0.8.1
six==1.9.0
software-center-aptd-plugins==0.0.0
sqlparse==0.1.13
system-service==0.1.6
traitlets==4.2.1
Twisted-Core==14.0.2
Twisted-Web==14.0.2
unity-lens-photos==1.0
urllib3==1.9.1
virtualenv==1.11.6
xdiagnose==3.7.2
zc.buildout==1.7.1
zope.interface==4.1.1
